requirements.txt
appdirs==1.4.4
dnspython==1.16.0
filelock==3.0.12
idna==3.2
irc==19.0.1
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.3.0
jaraco.logging==3.1.0
jaraco.stream==3.0.2
jaraco.text==3.5.1
more-itertools==8.9.0
pymongo==3.12.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
tempora==4.1.1
uritools==3.0.2
urlextract==1.3.0
